Created Product Details page
Created the page using swift with iOS dependency, SwiftUI libraries. This page acts as a "middle man" when a user would like to add products to their shopping list, when searched for. This page shows the user a details about the product in question, allowing them to see the product name, product image, nutrition details about the product, the ingredients list, the price of the product, and an option to add the product to cart or go back.