Created Index
Created the page using Swift with iOS dependency, SwiftUI libraries. Primarily used to test the placement of visual assets of the application, this page contains the gradient function, creating a linear gradient and applying it as a background (using stacks, where ZStack places assets on the background, HStack places items horizontally adjacent, VStack vertically stacks items, similar to that of a list.